Rename getDirectoryInstanceIdByDirectoryPath to match canonical export
